Move function to util
WIP
